Move registerPlugins to cache as a part of its API
This change adds more natural way of applying plugins on cache instance.
Right now you will be able to register plugins on the cache object
instead of having to create a new object each time you register them (plugins).

What about different plugins for different instances?
Simply, make a copy of a cache object or create a new instance.
It's now up to the user how (s)he will go about it.
