+++ src/index.js
@@ -1,11 +1,9 @@
 import { createCache, getPreData, getPostData } from './createCache';
 import createItem from './createItem';
-import registerPlugins from './registerPlugins';
 
 export {
     createCache,
     createItem,
     getPreData,
-    getPostData,
-    registerPlugins
+    getPostData
 };
